+++ src/Miningcore/Blockchain/Bitcoin/BitcoinJobManager.cs
@@ -33,7 +33,7 @@ public class BitcoinJobManager : BitcoinJobManagerBase<BitcoinJob>
         this.blockCandidateRecorder = blockCandidateRecorder;
     }
 
-    private BitcoinTemplate coin;
+    protected BitcoinTemplate coin;
     private readonly IBlockCandidateRecorder blockCandidateRecorder;
     private bool bip54CoinbaseEnabled;
     private bool directCoinbasePayoutEnabled;
@@ -120,9 +120,20 @@ public class BitcoinJobManager : BitcoinJobManagerBase<BitcoinJob>
         return new RpcResponse<BlockTemplate>(result!.ResultAs<BlockTemplate>());
     }
 
-    private BitcoinJob CreateJob()
+    protected virtual BitcoinJob CreateJob()
     {
         return new();
+    }
+
+    protected virtual void InitializeJob(BitcoinJob job,
+        BlockTemplate blockTemplate)
+    {
+        job.InitWithCoinbasePolicy(blockTemplate, NextJobId(),
+            poolConfig, extraPoolConfig, clusterConfig, clock,
+            poolAddressDestination, network, isPoS, ShareMultiplier,
+            coin.CoinbaseHasherValue, coin.HeaderHasherValue,
+            !isPoS ? coin.BlockHasherValue : coin.PoSBlockHasherValue ??
+                coin.BlockHasherValue, bip54CoinbaseEnabled);
     }
 
     protected override void PostChainIdentifyConfigure()
@@ -217,11 +228,7 @@ public class BitcoinJobManager : BitcoinJobManagerBase<BitcoinJob>
 
                 job = CreateJob();
 
-                job.InitWithCoinbasePolicy(blockTemplate, NextJobId(),
-                    poolConfig, extraPoolConfig, clusterConfig, clock, poolAddressDestination, network, isPoS,
-                    ShareMultiplier, coin.CoinbaseHasherValue, coin.HeaderHasherValue,
-                    !isPoS ? coin.BlockHasherValue : coin.PoSBlockHasherValue ?? coin.BlockHasherValue,
-                    bip54CoinbaseEnabled);
+                InitializeJob(job, blockTemplate);
 
                 if(isNew)
                 {